Refactor authentication and route registration to use primitive.ObjectID and config struct
- Updated auth.go to use primitive.ObjectID for user ID handling and JWT generation.
- Modified document.go, finance.go, schedule.go, user.go, and routes.go to accept config.Config parameter in route registration functions.
- Enhanced middleware calls to include configuration in route handlers.
- Updated go.mod to use newer versions of dependencies and added indirect dependencies.
- Fixed password verification to use uint8 for threads in argon2 hashing.

+++ backend/utils/password.go
@@ -48,7 +48,7 @@ func VerifyPassword(password, encodedHash string) (bool, error) {
 	if err != nil {
 		return false, err
 	}
-	calcHash := argon2.IDKey([]byte(password), salt, time, memory, threads, uint32(len(hash)))
+	calcHash := argon2.IDKey([]byte(password), salt, time, memory, uint8(threads), uint32(len(hash)))
 	if !compareHash(hash, calcHash) {
 		return false, nil
 	}
